feat: create function list all orders and update in price order-product

+++ src/migration/1744799963484-create-table-order-product.ts
@@ -10,7 +10,7 @@ import { MigrationInterface, QueryRunner } from 'typeorm';
                  order_id int NOT NULL,
                  product_id int NOT NULL,
                  amount int NOT NULL,
-                 price int NOT NULL,
+                 price double precision NOT NULL,
                  created_at timestamp without time zone DEFAULT now() NOT NULL,
                  updated_at timestamp without time zone DEFAULT now() NOT NULL,
                  primary key (id),
